<Feat> :  이동 수단 선택 ---------------------------------------- 1. 여행 일정 페이지 분리 2. 이동 수단 및 예상 소요 시간
----------------------------------------
1. 앞으로 할 일
